chore: Refactor routes to propertyRoute to pave way for boards
